+++ app/Services/AiAssistant/Exceptions/AiProviderException.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Services\AiAssistant\Exceptions;
+
+use RuntimeException;
+
+class AiProviderException extends RuntimeException
+{
+}